+++ CHANGELOG.md
@@ -19,6 +19,7 @@
     - `math::GridAab::intersection_box()` is a new definition of box intersection which returns a bigger box in some zero-volume/flat cases.
     - `math::GridAab::union_cubes()` is a new definition of union which returns a smaller box in zero-volume/flat cases.
     - `math::GridSize` type alias for `euclid::Size3D<GridCoordinate, Cube>`.
+    - `math::Vol::subdivide()` splits a `Vol`'s bounds and data into parts that can be processed in parallel.
 
     - `transaction::Transactional::transact()` provides a convenient way to create and immediately execute transactions.